feat: Chat 화면 라우터 적용
